desktop,web: refactor mods, assets, etc.
- web is broken ATM, need to refactor deps for desktop and web
  (specifically tokio, sqlx)
