+++ Cargo.toml
@@ -1,13 +1,13 @@
 [workspace]
 resolver = "2"
-members = ["api", "boil", "core", "desktop", "web"]
+members = ["api", "boil", "core", "desktop", "views", "web"]
 
 [workspace.dependencies]
 async-trait = "0.1.89"
 chrono = "0.4.42"
 clap = { version = "4.5.47", features = ["derive"] }
 color-eyre = "0.6.5"
-colored = "3.0.0" # TODO: pick one of colored or nu-ansi-term
+colored = "3.0.0"                                                             # TODO: pick one of colored or nu-ansi-term
 dioxus = { version = "0.6.3" }
 fs_extra = "1.3.0"
 dirs = "6.0.0"
@@ -21,7 +21,8 @@ serde = { version = "1.0.225", features = ["derive"] }
 serde_json = "1.0.145"
 sqlx = { version = "0.8", features = ["runtime-tokio", "sqlite", "migrate"] }
 tabled = "0.20.0"
-tokio = { version = "1.47.1", features = ["full"] }
+#tokio = { version = "1.47.1", features = ["full"] }
+tokio = { version = "1", default-features = false, features = ["rt", "macros"] }
 toml = "0.9.7"
 tracing = { version = "0.1.41", features = ["log"] }
 tracing-error = "0.2.1"
@@ -33,11 +34,13 @@ tracing-subscriber = { version = "0.3.20", features = [
 uuid = { version = "1.18.1", features = ["v4", "serde"] }
 walkdir = "2.5.0"
 
+
 # workspace
 api = { path = "api" }
 boil = { path = "boil" }
 desktop = { path = "desktop" }
 core = { path = "core" }
+views = { path = "views" }
 web = { path = "web" }
 
 [profile]
