+++ views/src/lib.rs
@@ -1,0 +1,13 @@
+mod constants;
+mod docs;
+mod echo;
+mod get_involved;
+mod search;
+mod templates;
+
+pub use constants::*;
+pub use docs::*;
+pub use echo::*;
+pub use get_involved::*;
+pub use search::*;
+pub use templates::*;
