+++ web/src/ui/mod.rs
@@ -1,0 +1,2 @@
+mod home;
+pub use home::Home;
