--- web/src/views/constants.rs
@@ -1,11 +1,0 @@
-use dioxus::prelude::*;
-
-pub const FAVICON: Asset = asset!("/assets/logo-flame.png");
-pub const MAIN_CSS: Asset = asset!("/assets/main.css");
-pub const TAILWIND_CSS: Asset = asset!("/assets/tailwind.css");
-
-pub const HIGHLIGHT_JS: Asset = asset!("/assets/highlight.min.js");
-pub const GITHUB_LIGHT_CSS: Asset = asset!("/assets/github.min.css");
-pub const GITHUB_DARK_CSS: Asset = asset!("/assets/github-dark.min.css");
-
-pub const LINK_STYLE: &str = "text-blue-400 px-1";
