--- web/src/views/echo.rs
@@ -1,31 +1,0 @@
-use api;
-use dioxus::prelude::*; // Import the api module from your crate (adjust the path if needed)
-
-// const ECHO_CSS: Asset = asset!("/assets/styling/echo.css");
-
-/// Echo component that demonstrates fullstack server functions.
-#[component]
-pub fn Echo() -> Element {
-    let mut response = use_signal(|| String::new());
-
-    rsx! {
-        // document::Link { rel: "stylesheet", href: ECHO_CSS }
-        div { id: "echo",
-            h4 { "ServerFn Echo" }
-            input {
-                placeholder: "Type here to echo...",
-                oninput: move |event| async move {
-                    let data = api::echo(event.value()).await.unwrap();
-                    response.set(data);
-                },
-            }
-
-            if !response().is_empty() {
-                p {
-                    "Server echoed: "
-                    i { "{response}" }
-                }
-            }
-        }
-    }
-}
