--- web/src/views/mod.rs
@@ -1,14 +1,0 @@
-pub mod constants;
-mod docs;
-mod echo;
-mod get_involved;
-mod home;
-mod search;
-mod templates;
-
-pub use docs::Docs;
-pub use echo::Echo;
-pub use get_involved::GetInvolved;
-pub use home::Home;
-pub use search::Search;
-pub use templates::Templates;
